Fix wring variable for time count id type Integer > Long

+++ app/src/main/java/org/biologer/biologer/gui/ActivityTimedCount.java
@@ -497,7 +497,7 @@ public class ActivityTimedCount extends AppCompatActivity implements FragmentTim
                 false,
                 false,
                 taxon.getId(),
-                viewModel.getServerId().intValue(),
+                viewModel.getServerId(),
                 taxon.getLatinName(),
                 String.valueOf(year),
                 String.valueOf(month),
